feat: Support creating bootable USB drives from unpartitioned USB disks
- Add automatic partition creation for unpartitioned USB disks using fdisk
- Optimize df command to exclude network filesystems (NFS, CIFS, SMBFS, SSHFS, FTPFS, DAVFS)
- Improve device scanning logic to skip network-related devices

Log: Enable creating bootable USB drives from unpartitioned USB disks and improve disk querying efficiency

Bug: https://pms.uniontech.com/bug-view-338629.html

+++ src/libdbm/util/deviceinfo.h
@@ -12,14 +12,14 @@ class DeviceInfo
 {
 public:
     DeviceInfo() {}
-    DeviceInfo(const QString &path, quint32 used, quint32 total, const QString &label)
+    DeviceInfo(const QString &path, quint64 used, quint64 total, const QString &label)
         : path(path),  used(used), total(total), label(label)
     {
     }
 
     QString path    = "";
-    quint32 used    = 0;
-    quint32 total   = 0;
+    quint64 used    = 0;
+    quint64 total   = 0;
     QString label   = QObject::tr("Removable Disk");
     QString uuid    = "";
     QString fstype  = "";
